backend/requirements.txt
Flask
gunicorn
requests
PyYAML
beautifulsoup4
